+++ app/Http/Controllers/Auth/UserController.php
@@ -375,18 +375,18 @@ class UserController extends Controller
 
     /**
      * @OA\Post (
-     *     path="/api/user/reset-password",
-     *     tags={"학생"},
-     *     summary="비밀번호 초기화",
-     *     description="회원가입 시 입력한 이름, 이메일을 검증하고, 메일 전송 후 코드를 인증",
+     *     path="/api/find-email",
+     *     tags={"유저"},
+     *     summary="이메일 찾기",
+     *     description="회원가입 시 입력한 이름과 전화번호를 통하여 일치하는 값을 가진 이메일을 찾음",
      *     @OA\RequestBody(
-     *         description="name & email",
+     *         description="name & phone_number(without hyphen)",
      *         required=true,
      *         @OA\MediaType(
      *             mediaType="application/json",
      *             @OA\Schema (
      *                 @OA\Property (property="name", type="string", description="이름", example="testname"),
-     *                 @OA\Property (property="email", type="string", description="이메일", example="test@test.com"),
+     *                 @OA\Property (property="phone_number", type="string", description="휴대폰 번호", example="01012345678"),
      *             )
      *         )
      *     ),
@@ -396,69 +396,24 @@ class UserController extends Controller
      *     @OA\Response(response="500", description="ServerError"),
      * )
      */
-    public function resetPassword(Request $request): \Illuminate\Http\JsonResponse
+    public function findEmail(Request $request): \Illuminate\Http\JsonResponse
     {
         try {
             $validated = $request->validate([
-                'email' => 'required|email|exists:users,email',
-                'name'  => 'required|string',
+                'name' => 'required|string',
+                'phone_number' => 'required|string',
             ]);
-        } catch (ValidationException $exception) {
-            $errorStatus = $exception->status;
-            $errorMessage = $exception->getMessage();
-            return response()->json(['error' => $errorMessage], $errorStatus);
+        } catch (ValidationException $validationException) {
+            $errorStatus = $validationException->status;
+            $errorMessage = $validationException->getMessage();
+            return response()->json(['error'=>$errorMessage], $errorStatus);
         }
 
         try {
-            User::where('email', $validated['email'])->where('name', $validated['name'])->firstOrFail();
+            $admin = User::where('phone_number', $validated['phone_number'])->where('name', $validated['name'])->firstOrFail();
         } catch (ModelNotFoundException) {
-            return response()->json(['error' => '해당하는 유저가 존재하지 않습니다.'], 404);
+            return response()->json(['error' => $this->modelExceptionMessage], 404);
         }
-
-        $resetPasswordService = new ResetPasswordService($validated['email']);
-
-        return $resetPasswordService();
-    }
-
-    /**
-     * @OA\Patch (
-     *     path="/api/user/password",
-     *     tags={"학생"},
-     *     summary="비밀번호 재설정",
-     *     description="유저의 비밀번호 재설정 시 사용합니다.",
-     *     @OA\Requestbody(
-     *         description="수정할 유저의 비밀번호",
-     *         required=true,
-     *         @OA\Mediatype(
-     *             mediaType="application/json",
-     *             @OA\Schema (
-     *                  @OA\Property (property="password", type="string", description="변경할 비밀번호", example="asdf123"),
-     *             )
-     *         ),
-     *     ),
-     *     @OA\Response(response="200", description="Success"),
-     *     @OA\Response(response="404", description="ModelNotFoundException"),
-     *     @OA\Response(response="422", description="ValidationException"),
-     *     @OA\Response(response="500", description="Server Error"),
-     * )
-     */
-    public function recoverPassword(Request $request): \Illuminate\Http\JsonResponse
-    {
-        try {
-            $validated = $request->validate([
-                'password' => 'required|string',
-            ]);
-        } catch (ValidationException $exception) {
-            $errorStatus = $exception->status;
-            $errorMessage = $exception->getMessage();
-            return response()->json(['error' => $errorMessage], $errorStatus);
-        }
-
-        $user = auth()->user();
-        $user->password = Hash::make($validated['password']);
-
-        if(!$user->save()) return response()->json(['error' => '비밀번호 변경에 실패하였습니다.'], 500);
-
-        return response()->json(['message' => '비밀번호가 변경되었습니다.']);
+        return response()->json(['admin' => $admin]);
     }
 }
